# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Reciclado de Alimentos - API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List available products
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: lat
          in: query
          schema:
            type: ["number", "null"]
        - name: lng
          in: query
          schema:
            type: ["number", "null"]
        - name: radius
          in: query
          schema:
            type: ["number", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (merchant only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (merchant only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (merchant only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MerchantProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        legalName:
          type: ["string", "null"]
        cuit:
          type: ["string", "null"]
        category:
          type: string
        description:
          type: ["string", "null"]
        phone:
          type: string
        addressLine:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        pickupHours:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isOpen:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - businessName
        - category
        - phone
        - addressLine
        - city
        - province
        - isVerified
        - isOpen
        - createdAt
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        icon:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - id
        - name
        - slug
        - isActive
    Product:
      type: object
      properties:
        id:
          type: string
        merchantId:
          type: string
        categoryId:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        originalPrice:
          type: number
        salePrice:
          type: number
        discountPercent:
          type: number
        quantityAvailable:
          type: integer
        unit:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        pickupDeadline:
          type: ["string", "null"]
        pickupAddress:
          type: ["string", "null"]
        pickupLatitude:
          type: ["number", "null"]
        pickupLongitude:
          type: ["number", "null"]
        status:
          type: string
        isFeatured:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        merchant:
          $ref: "#/components/schemas/MerchantProfile"
        category:
          $ref: "#/components/schemas/Category"
      required:
        - id
        - merchantId
        - categoryId
        - name
        - slug
        - originalPrice
        - salePrice
        - discountPercent
        - quantityAvailable
        - status
        - isFeatured
        - createdAt
        - merchant
        - category
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - products
        - total
        - page
        - limit
...
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: string
        imageUrl:
          type: ["string", "null"]
        originalPrice:
          type: number
        salePrice:
          type: number
        quantityAvailable:
          type: integer
        unit:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        pickupDeadline:
          type: ["string", "null"]
        pickupAddress:
          type: ["string", "null"]
        pickupLatitude:
          type: ["number", "null"]
        pickupLongitude:
          type: ["number", "null"]
        isFeatured:
          type: boolean
        status:
          type: string
      required:
        - name
        - categoryId
        - originalPrice
        - salePrice
        - quantityAvailable
    UpdateProductBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        categoryId:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        originalPrice:
          type: ["number", "null"]
        salePrice:
          type: ["number", "null"]
        quantityAvailable:
          type: ["integer", "null"]
        unit:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        pickupDeadline:
          type: ["string", "null"]
        pickupAddress:
          type: ["string", "null"]
        pickupLatitude:
          type: ["number", "null"]
        pickupLongitude:
          type: ["number", "null"]
        isFeatured:
          type: ["boolean", "null"]
        status:
          type: ["string", "null"]
...