# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reciclado de Alimentos - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  /** @nullable */
  description?: string | null;
  categoryId: string;
  /** @nullable */
  imageUrl?: string | null;
  originalPrice: number;
  salePrice: number;
  quantityAvailable: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  pickupDeadline?: string | null;
  /** @nullable */
  pickupAddress?: string | null;
  /** @nullable */
  pickupLatitude?: number | null;
  /** @nullable */
  pickupLongitude?: number | null;
  isFeatured?: boolean;
  status?: string;
}